requirements.txt
python-telegram-bot[job-queue]~=21.0
gspread~=6.1
google-auth~=2.29
python-dotenv~=1.0
